LayoutTests/accessibility/display-contents/table-dynamic-expected.txt

This test ensures that display:contents tables with dynamic content are accessible.

Dumping table #t0 (expecting row count 3, column count 2)
PASS: table.rowCount === 3
PASS: table.columnCount === 2
PASS: table.cellForColumnAndRow(0, 0).domIdentifier === "r0c0-t0"
PASS: table.cellForColumnAndRow(1, 0).domIdentifier === "r0c1-t0"
PASS: table.cellForColumnAndRow(0, 1).domIdentifier === "r1c0-t0"
PASS: table.cellForColumnAndRow(1, 1).domIdentifier === "r1c1-t0"
PASS: table.cellForColumnAndRow(0, 2).domIdentifier === "r2c0-t0"
PASS: table.cellForColumnAndRow(1, 2).domIdentifier === "r2c1-t0"
Dumping table #t1 (expecting row count 3, column count 2)
PASS: table.rowCount === 3
PASS: table.columnCount === 2
PASS: table.cellForColumnAndRow(0, 0).domIdentifier === "r0c0-t1"
PASS: table.cellForColumnAndRow(1, 0).domIdentifier === "r0c1-t1"
PASS: table.cellForColumnAndRow(0, 1).domIdentifier === "r1c0-t1"
PASS: table.cellForColumnAndRow(1, 1).domIdentifier === "r1c1-t1"
PASS: table.cellForColumnAndRow(0, 2).domIdentifier === "r2c0-t1"
PASS: table.cellForColumnAndRow(1, 2).domIdentifier === "r2c1-t1"

Performing search traversal of body.

{#t0 AXRole: AXTable} (parent: {#body AXRole: AXGroup})

{#r0-t0 AXRole: AXRow} (parent: {#t0 AXRole: AXTable})

{#r0c0-t0 AXRole: AXCell} (parent: {#r0-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Author} (parent: {#r0c0-t0 AXRole: AXCell})

{#r0c1-t0 AXRole: AXCell} (parent: {#r0-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Title} (parent: {#r0c1-t0 AXRole: AXCell})

{#r1-t0 AXRole: AXRow} (parent: {#t0 AXRole: AXTable})

{#r1c0-t0 AXRole: AXCell} (parent: {#r1-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Stephen Hawking} (parent: {#r1c0-t0 AXRole: AXCell})

{#r1c1-t0 AXRole: AXCell} (parent: {#r1-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) A Brief History of Time} (parent: {#r1c1-t0 AXRole: AXCell})

{#r2-t0 AXRole: AXRow} (parent: {#t0 AXRole: AXTable})

{#r2c0-t0 AXRole: AXCell} (parent: {#r2-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Carl Sagan} (parent: {#r2c0-t0 AXRole: AXCell})

{#r2c1-t0 AXRole: AXCell} (parent: {#r2-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Cosmos} (parent: {#r2c1-t0 AXRole: AXCell})

{#t1 AXRole: AXTable} (parent: {#body AXRole: AXGroup})

{#r0-t1 AXRole: AXRow} (parent: {#t1 AXRole: AXTable})

{#r0c0-t1 AXRole: AXCell} (parent: {#r0-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Author} (parent: {#r0c0-t1 AXRole: AXCell})

{#r0c1-t1 AXRole: AXCell} (parent: {#r0-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Title} (parent: {#r0c1-t1 AXRole: AXCell})

{#r1-t1 AXRole: AXRow} (parent: {#t1 AXRole: AXTable})

{#r1c0-t1 AXRole: AXCell} (parent: {#r1-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Stephen Hawking} (parent: {#r1c0-t1 AXRole: AXCell})

{#r1c1-t1 AXRole: AXCell} (parent: {#r1-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) A Brief History of Time} (parent: {#r1c1-t1 AXRole: AXCell})

{#r2-t1 AXRole: AXRow} (parent: {#t1 AXRole: AXTable})

{#r2c0-t1 AXRole: AXCell} (parent: {#r2-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Carl Sagan} (parent: {#r2c0-t1 AXRole: AXCell})

{#r2c1-t1 AXRole: AXCell} (parent: {#r2-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Cosmos} (parent: {#r2c1-t1 AXRole: AXCell})

Moving #r2-t1 from table one to table zero.
Dumping table #t0 (expecting row count 4, column count 2)
PASS: table.rowCount === 4
PASS: table.columnCount === 2
table.cellForColumnAndRow(0, 0).domIdentifier is r0c0-t0
table.cellForColumnAndRow(1, 0).domIdentifier is r0c1-t0
table.cellForColumnAndRow(0, 1).domIdentifier is r1c0-t0
table.cellForColumnAndRow(1, 1).domIdentifier is r1c1-t0
table.cellForColumnAndRow(0, 2).domIdentifier is r2c0-t0
table.cellForColumnAndRow(1, 2).domIdentifier is r2c1-t0
table.cellForColumnAndRow(0, 3).domIdentifier is r2c0-t1
table.cellForColumnAndRow(1, 3).domIdentifier is r2c1-t1
Dumping table #t1 (expecting row count 2, column count 2)
PASS: table.rowCount === 2
PASS: table.columnCount === 2
PASS: table.cellForColumnAndRow(0, 0).domIdentifier === "r0c0-t1"
PASS: table.cellForColumnAndRow(1, 0).domIdentifier === "r0c1-t1"
PASS: table.cellForColumnAndRow(0, 1).domIdentifier === "r1c0-t1"
PASS: table.cellForColumnAndRow(1, 1).domIdentifier === "r1c1-t1"

Performing search traversal of body.

{#t0 AXRole: AXTable} (parent: {#body AXRole: AXGroup})

{#r0-t0 AXRole: AXRow} (parent: {#t0 AXRole: AXTable})

{#r0c0-t0 AXRole: AXCell} (parent: {#r0-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Author} (parent: {#r0c0-t0 AXRole: AXCell})

{#r0c1-t0 AXRole: AXCell} (parent: {#r0-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Title} (parent: {#r0c1-t0 AXRole: AXCell})

{#r1-t0 AXRole: AXRow} (parent: {#t0 AXRole: AXTable})

{#r1c0-t0 AXRole: AXCell} (parent: {#r1-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Stephen Hawking} (parent: {#r1c0-t0 AXRole: AXCell})

{#r1c1-t0 AXRole: AXCell} (parent: {#r1-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) A Brief History of Time} (parent: {#r1c1-t0 AXRole: AXCell})

{#r2-t0 AXRole: AXRow} (parent: {#t0 AXRole: AXTable})

{#r2c0-t0 AXRole: AXCell} (parent: {#r2-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Carl Sagan} (parent: {#r2c0-t0 AXRole: AXCell})

{#r2c1-t0 AXRole: AXCell} (parent: {#r2-t0 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T0) Cosmos} (parent: {#r2c1-t0 AXRole: AXCell})

{#r2-t1 AXRole: AXRow} (parent: {#t0 AXRole: AXTable})

{#r2c0-t1 AXRole: AXCell} (parent: {#r2-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Carl Sagan} (parent: {#r2c0-t1 AXRole: AXCell})

{#r2c1-t1 AXRole: AXCell} (parent: {#r2-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Cosmos} (parent: {#r2c1-t1 AXRole: AXCell})

{#t1 AXRole: AXTable} (parent: {#body AXRole: AXGroup})

{#r0-t1 AXRole: AXRow} (parent: {#t1 AXRole: AXTable})

{#r0c0-t1 AXRole: AXCell} (parent: {#r0-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Author} (parent: {#r0c0-t1 AXRole: AXCell})

{#r0c1-t1 AXRole: AXCell} (parent: {#r0-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Title} (parent: {#r0c1-t1 AXRole: AXCell})

{#r1-t1 AXRole: AXRow} (parent: {#t1 AXRole: AXTable})

{#r1c0-t1 AXRole: AXCell} (parent: {#r1-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) Stephen Hawking} (parent: {#r1c0-t1 AXRole: AXCell})

{#r1c1-t1 AXRole: AXCell} (parent: {#r1-t1 AXRole: AXRow})

{AXRole: AXStaticText AXValue: (T1) A Brief History of Time} (parent: {#r1c1-t1 AXRole: AXCell})

PASS successfullyParsed is true

TEST COMPLETE
Table zero caption
(T0) Author (T0) Title
(T0) Stephen Hawking (T0) A Brief History of Time
(T0) Carl Sagan (T0) Cosmos
(T1) Carl Sagan (T1) Cosmos
Table one caption
(T1) Author (T1) Title
(T1) Stephen Hawking (T1) A Brief History of Time
